Admin/user can update and detele his own post with laravel policies

+++ app/Http/Controllers/PostController.php
@@ -13,6 +13,8 @@ class PostController extends Controller
 
         $posts = Post::all();
 
+        //$posts = auth()->user()->posts;
+
         return view('admin.posts.index', ['posts'=> $posts]);
 
     }
@@ -25,13 +27,21 @@ class PostController extends Controller
         return view('blog-post', ['post'=> $post]);
     }
 
+
+
     public function create(){
         return view('admin.posts.create');
     }
 
+
+
     public function edit(Post $post){
 
         //$posts = Post::findOrFail($post);
+
+        // $this->authorize('view',$post);
+
+        //if(auth()->user()->can('view',$post)){}
 
         return view('admin.posts.edit', ['post'=> $post]);
 
@@ -50,6 +60,8 @@ class PostController extends Controller
         $post->body = $inputs['body'];
 
         //auth()->user()->posts()->save($post);
+
+        $this->authorize('update',$post);
 
         $post->save();
         //$post->update();
@@ -98,6 +110,9 @@ class PostController extends Controller
 
     public function destroy(Post $post){
 
+        //if(auth()->user()->id != $post->user_id){}
+
+        $this->authorize('delete',$post);
         $post->delete();
 
         //Session::flash('message','Post is deleted');
